sequential finally done

+++ order.hpp
@@ -13,7 +13,7 @@ struct Order
     uint32_t order_id; 
     uint32_t price; 
     uint32_t count; 
-    uint32_t time;
+    intmax_t time;
     std::string instrument; 
     uint32_t execution_id;
